getting unit tests to work
