+++ api/app/tests/hfi/test_hfi_endpoint_request.py
@@ -9,10 +9,10 @@ from aiohttp import ClientSession
 from pytest_mock import MockerFixture
 import app.main
 import app.routers.hfi_calc
-from app.db.models.hfi_calc import PlanningWeatherStation, FuelType, FireCentre, PlanningArea, HFIRequest
+from app.db.models.hfi_calc import (PlanningWeatherStation, FuelType, FireCentre, PlanningArea,
+                                    HFIRequest, FireStartRange, FireStartLookup)
 from app.tests.common import default_mock_client_get
-from app.tests import load_json_file_with_name, load_json_file
-from app.tests.hfi import mock_station_crud
+from app.tests import load_json_file
 
 
 def _setup_mock(monkeypatch: pytest.MonkeyPatch):
@@ -38,10 +38,93 @@ def _setup_mock(monkeypatch: pytest.MonkeyPatch):
                                     planning_area_id=2), fuel_type_2, planning_area_2, fire_centre)
         )
 
-    monkeypatch.setattr(app.hfi.hfi_calc, 'get_fire_weather_stations', mock_get_fire_weather_stations)
+    code1 = 230
+    code2 = 239
+    all_station_codes = [{'station_code': code1}, {'station_code': code2}]
 
-    # mock out database calls:
-    mock_station_crud(monkeypatch)
+    def mock_get_all_stations(__):
+        """ Returns mocked WFWXWeatherStations codes. """
+        return all_station_codes
+
+    def mock_get_fire_centre_stations(_, __: int):
+        """ Returns mocked WFWXWeatherStation with fuel types. """
+        def get_fuel_type_code_by_station_code(code: int):
+            if code == 230:
+                return 'C3'
+            return 'C7B'
+        result = []
+        for station_code, planning_area_id in [(230, 1), (239, 1), (230, 2)]:
+            planning_station = PlanningWeatherStation(
+                station_code=station_code, planning_area_id=planning_area_id)
+            fuel_type_code = get_fuel_type_code_by_station_code(station_code)
+            fuel_type = FuelType(id=1, abbrev=fuel_type_code, fuel_type_code=fuel_type_code,
+                                 description=fuel_type_code,
+                                 percentage_conifer=100, percentage_dead_fir=0)
+            result.append((planning_station, fuel_type))
+        return result
+
+    def mock_get_fire_centre_fire_start_ranges(_, __: int):
+        """ Returns mocked FireStartRange """
+        data = ((1, '0-1'), (2, '1-2'), (3, '2-3'), (4, '3-6'), (5, '6+'))
+        return [FireStartRange(id=id, label=range) for id, range in data]
+
+    def mock_get_fire_start_lookup(_):
+        """ Returns mocked FireStartLookup """
+        data = ((1, 1, 1, 1),
+                (2, 1, 2, 1),
+                (3, 1, 3, 2),
+                (4, 1, 4, 3),
+                (5, 1, 5, 4),
+                (6, 2, 1, 1),
+                (7, 2, 2, 1),
+                (8, 2, 3, 2),
+                (9, 2, 4, 4),
+                (10, 2, 5, 5),
+                (11, 3, 1, 2),
+                (12, 3, 2, 3),
+                (13, 3, 3, 4),
+                (14, 3, 4, 5),
+                (15, 3, 5, 6),
+                (16, 4, 1, 3),
+                (17, 4, 2, 4),
+                (18, 4, 3, 4),
+                (19, 4, 4, 5),
+                (20, 4, 5, 6),
+                (21, 5, 1, 4),
+                (22, 5, 2, 5),
+                (23, 5, 3, 6),
+                (24, 5, 4, 6),
+                (25, 5, 5, 6))
+        return [FireStartLookup(id=id,
+                                fire_start_range_id=fire_start_range_id,
+                                mean_intensity_group=mean_intensity_group,
+                                prep_level=prep_level) for
+                id, fire_start_range_id, mean_intensity_group, prep_level in data]
+
+    fuel_types = [
+        FuelType(id=1, abbrev='O1B', fuel_type_code='O1B', description='O1B',
+                 percentage_conifer=0, percentage_dead_fir=0),
+        FuelType(id=1, abbrev='C7B', fuel_type_code='C7B', description='C7B',
+                 percentage_conifer=100, percentage_dead_fir=0)
+    ]
+
+    def mock_get_fuel_type_by_id(_, fuel_type_id: int):
+        """ Returns mocked FuelType """
+        return next(fuel_type for fuel_type in fuel_types if fuel_type.id == fuel_type_id)
+
+    def mock_get_fuel_types(_):
+        return fuel_types
+
+    monkeypatch.setattr(app.hfi.hfi_calc, 'get_fire_weather_stations', mock_get_fire_weather_stations)
+    monkeypatch.setattr(app.db.crud.hfi_calc, 'get_all_stations', mock_get_all_stations)
+    # TODO: this is problematic, why are we calling get_fire_centre_stations twice?
+    monkeypatch.setattr(app.hfi.hfi_calc, 'get_fire_centre_stations', mock_get_fire_centre_stations)
+    monkeypatch.setattr(app.hfi.hfi_calc, 'get_fire_centre_fire_start_ranges',
+                        mock_get_fire_centre_fire_start_ranges)
+    monkeypatch.setattr(app.hfi.hfi_calc, 'get_fuel_types', mock_get_fuel_types)
+    monkeypatch.setattr(app.hfi.hfi_calc, 'get_fire_start_lookup', mock_get_fire_start_lookup)
+    monkeypatch.setattr(app.routers.hfi_calc, 'get_fire_centre_stations', mock_get_fire_centre_stations)
+    monkeypatch.setattr(app.routers.hfi_calc, 'get_fuel_type_by_id', mock_get_fuel_type_by_id)
 
 
 @pytest.mark.usefixtures('mock_jwt_decode')
